requirements.txt
addict==2.4.0
apex==0.9.10dev
colorlog==6.7.0
einops==0.7.0
evaluate==0.4.1
fvcore==0.1.5.post20221221
h5py==3.9.0
imageio==2.31.1
joblib==1.2.0
lmdb==1.4.1
matplotlib==3.7.1
mc==0.0
mock==5.1.0
natsort==8.4.0
nni==3.0
numpy==1.24.3
packaging==23.2
pandas==2.0.3
Pillow==9.3.0
Pillow==10.1.0
psutil==5.9.5
ptflops==0.7.1.2
pytest==7.4.3
pytorch_sphinx_theme==0.0.19
PyYAML==6.0
PyYAML==6.0.1
regex==2023.5.5
scipy==1.11.4
setuptools==68.0.0
skimage==0.0
Sphinx==7.2.6
tensorboardX==2.6.2.2
tensorflow==2.13.0
tensorflow_intel==2.13.0
termcolor==2.4.0
tifffile==2023.8.30
timm==0.9.2
torch==2.1.0+cu121
torch_tb_profiler==0.4.3
torchvision==0.16.0
tqdm==4.65.0
turbojpeg==0.0.2
yapf==0.40.1
